Moving sorting functions to their own file
